refactor: abstract graphql queries and mutations to one file
