+++ src/components/Topic/Topic.tsx
@@ -25,8 +25,6 @@ const Topic: React.FC<TopicNodeProps> = ({
   }
   const { data } = useQuery(gql`{ selectedTopicTitle, selectedTopicId }`);
   return (
-    // TODO: "Coding for the keyboard is important for users with physical disabilities
-    // who cannot use a mouse, AT compatibility, and screenreader users."
     // eslint-disable-next-line jsx-a11y/click-events-have-key-events
     <div
       className={isPreview ? 'no-anim topic-container' : 'topic-container'}
