+++ src/containers/RoadmapDashboard/RoadmapDashboard.tsx
@@ -1,4 +1,3 @@
-import gql from 'graphql-tag';
 import jwtDecode from 'jwt-decode';
 import React from 'react';
 import { Redirect, RouteComponentProps } from 'react-router-dom'; // eslint-disable-line
@@ -7,20 +6,9 @@ import './RoadmapDashboard.css';
 import RoadmapTree from '../RoadmapTree/RoadmapTree';
 import TopicDetails from '../../components/TopicDetails/TopicDetails';
 import Navbar from '../../components/Navbar/Navbar';
+import { CHECK_ROADMAP_USER, GET_TOPIC_ID } from './RoadmapDashboard.Queries';
 
 type TParams = { id: string };
-
-const CHECK_ROADMAP_USER = gql`
-query roadmapUser($id: ID!) {
-  roadmaps(id: $id) {
-    UserId
-    title
-  }
-}`;
-
-const GET_TOPIC_ID = gql`{
-  selectedTopicId
-}`;
 
 // TODO: Review this component's type
 const RoadmapDashboard = ({ match }: RouteComponentProps<TParams>) => {
